Rework unit tests.
This commit does several things on the unit tests:

- upgrade QUnit 1.11 -> 1.20
- use qunit-cli instead of qunit (this let us use a recent version of QUnit)
- split the test/test.js file: we now have separated files in the
  test/asserts/ folder
- add the test files in the jshint list
- fix the indentation (3 spaces -> 4 spaces)
