+++ package.json
@@ -5,7 +5,7 @@
   "description": "Create, read and edit .zip files with Javascript http://stuartk.com/jszip",
   "scripts": {
     "test": "npm run test-node && npm run test-browser",
-    "test-node": "cd test && qunit -c node.js -t test.js",
+    "test-node": "qunit-cli -c test/helpers/test-utils.js test/helpers/node-test-utils.js test/asserts/*.js",
     "test-browser": "grunt build && grunt test",
     "lint": "grunt jshint"
   },
@@ -34,7 +34,6 @@
     "inflate"
   ],
   "devDependencies": {
-    "qunit": "^0.7.6",
     "grunt": "~0.4.1",
     "grunt-cli": "~0.1.9",
     "grunt-saucelabs": "~8.6.1",
@@ -44,7 +43,9 @@
     "grunt-browserify": "~4.0.1",
     "grunt-contrib-jshint": "~0.11.3",
     "grunt-contrib-uglify": "~0.9.2",
-    "jszip-utils": "~0.0.2"
+    "jszip-utils": "~0.0.2",
+    "qunit-cli": "~0.2.0",
+    "qunitjs": "~1.20.0"
   },
   "dependencies": {
     "es6-promise": "~2.1.0",
