--- test/browser-test-utils.js
@@ -1,5 +1,0 @@
-var JSZipTestUtils = {
-    loadZipFile : function (name, callback) {
-        JSZipUtils.getBinaryContent(name + "?_=" + ( new Date() ).getTime(), callback);
-    }
-};
