+++ test/helpers/browser-test-utils.js
@@ -1,0 +1,5 @@
+/* global JSZip,JSZipUtils,JSZipTestUtils */
+'use strict';
+JSZipTestUtils.loadZipFile = function (name, callback) {
+    JSZipUtils.getBinaryContent(name + "?_=" + ( new Date() ).getTime(), callback);
+};
