--- test/node.js
@@ -1,14 +1,0 @@
-'use strict';
-
-var fs = require("fs");
-
-global.JSZip = require("../lib/index");
-
-global.JSZipTestUtils = {
-    loadZipFile: function(name, callback) {
-        fs.readFile(name, "binary", callback);
-    }
-};
-process.on('uncaughtException', function(err) {
-      console.log('uncaughtException: ' + err, err.stack);
-});
